combineTo1AuthContext
