+++ src/auth/AuthContext.js
@@ -1,14 +1,29 @@
 import React, { createContext, useContext, useState } from "react";
-import { fakeAuthProvider } from "./auth";
-import { useLocation } from "react-router-dom";
 
 const Auth = createContext(null);
 
-// UseAuth - to call the context
+const fakeAuthProvider = {
+  isAuthenticated: false,
+  signin(callback) {
+    setTimeout(() => {
+      // Simulating an asynchronous login operation
+      fakeAuthProvider.isAuthenticated = true;
+      callback();
+    }, 100);
+  },
+  signout(callback) {
+    setTimeout(() => {
+      // Simulating an asynchronous logout operation
+      fakeAuthProvider.isAuthenticated = false;
+      callback();
+    }, 100);
+  },
+};
+
 const useAuth = () => {
   return useContext(Auth);
 };
-// I still dont know what this does! need to revisit UseContext tutorial
+
 const AuthProvider = ({ children }) => {
   const [user, setUser] = useState(null);
 
@@ -31,11 +46,9 @@ const AuthProvider = ({ children }) => {
   return <Auth.Provider value={value}>{children}</Auth.Provider>;
 };
 
-// Wrapper wateva this is
 const RequireAuth = ({ children }) => {
   const auth = useAuth();
-  const location = useLocation();
-  // Use auth and location as needed
+  // You can use location if needed
   return children;
 };
