--- src/auth/AuthProvider.js
@@ -1,26 +1,0 @@
-import React, { useState } from "react";
-import { fakeAuthProvider } from "./auth";
-import AuthContext from "./AuthContext";
-
-function AuthProvider({ children }) {
-  let [user, setUser] = useState(null);
-
-  let singin = (newUser, callback) => {
-    return fakeAuthProvider.signin(() => {
-      setUser(newUser);
-      callback();
-    });
-  };
-
-  let signout = (callback) => {
-    return fakeAuthProvider.signout(() => {
-      setUser(null);
-      callback();
-    });
-  };
-
-  let value = { user, singin, signout };
-  return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
-}
-
-export default AuthProvider;
